+++ packages/sdk-metrics/src/export/PeriodicExportingMetricReader.ts
@@ -12,8 +12,9 @@ import {
 import { MetricReader } from './MetricReader';
 import type { PushMetricExporter } from './MetricExporter';
 import { callWithTimeout, TimeoutError } from '../utils';
-import type { MetricProducer } from './MetricProducer';
 import { InstrumentType } from './MetricData';
+import type { MetricProducer } from './MetricProducer';
+import { OTEL_COMPONENT_TYPE_VALUE_PERIODIC_METRIC_READER } from '../semconv';
 
 export type PeriodicExportingMetricReaderOptions = {
   /**
@@ -75,6 +76,7 @@ export class PeriodicExportingMetricReader extends MetricReader {
       aggregationSelector: exporter.selectAggregation?.bind(exporter),
       aggregationTemporalitySelector:
         exporter.selectAggregationTemporality?.bind(exporter),
+      otelComponentType: OTEL_COMPONENT_TYPE_VALUE_PERIODIC_METRIC_READER,
       metricProducers,
       cardinalitySelector: (instrumentType: InstrumentType) => {
         const limits = {